+++ src/controllers/search.ts
@@ -1,4 +1,5 @@
 import { ParameterizedContext } from "koa";
+import { format } from "util";
 import { searchView } from "../views/search-view";
 import { AppContext, AppState } from "../types";
 import { searchWaste } from "../elastic";
@@ -21,6 +22,7 @@ export async function search(ctx: ParameterizedContext<AppState, AppContext>) {
     }
   });
 
+  ctx.state.title = format(ctx.i18n("Gdzie wyrzucić \"%s\"?"), searchQuery)
   ctx.state.headerQuery = searchQuery;
   ctx.state.styles.push('/css/search.css');