+++ src/views/template-view.ts
@@ -1,5 +1,5 @@
 import { fastMapJoin } from "../helpers/fast-map-join";
-import { stylesheet, script } from "../helpers/html";
+import { stylesheet, script, escape } from "../helpers/html";
 
 interface Props {
   title: string;
@@ -15,7 +15,7 @@ export const templateView = (props: Props) => `<!doctype html>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <title>${escape(props.title)}</title>
-    <meta name="description" content="${escape(props.description)}">
+    ${props.description ? `<meta name="description" content="${escape(props.description)}">` : ""}
     ${fastMapJoin(props.styles, (href => stylesheet(href)))}
   </head>
   <body>