+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## `XX XXX XXXX
+ - [x] Add title for welcome & search page
+ - [ ] Add more items(parse from ekosystem.wroc.pl)
+
 ## 26 Aug 2019
  - [x] Add simple styles for header & search results
  - [x] Fix `elasticsearch` config